refactor: Update Testimonials.vue component to improve client testimonials
This commit updates the Testimonials.vue component to improve the client testimonials displayed on the page. The quotes from different clients have been revised to provide more specific and impactful feedback about the benefits of using Servcy. By enhancing the testimonials, we aim to attract potential customers and showcase the value of our product.