refactor: Optimize image loading in Transformation.vue
This commit optimizes the image loading in the Transformation.vue component by replacing the <img> tags with <NuxtImg> components. Lazy loading is implemented for all images, improving the page load performance by only loading images when they are visible in the viewport. Additionally, the alt text is added for each image, enhancing accessibility. These changes enhance the user experience and reduce the initial load time of the page.